moved handleClick and handleMouseMove to new file
